Change iterator to peek at the first element by if the peek view is at the same position as the iterator.
This behaviour improves similarity with the core::iter::Peekable iterator.
Besides, it is unlikely you didn't wanted to shift the view to the first element
and keep it None. This is also unlike how iterators usually work in Rust; they usually return None when the iterator is done, not when it "hasn't started yet".

Also improves documentation.