Update the CI config to support the added cfg(doctest).
Do not run doctest on the pinned version as it doesn't support cfg(doctest) on stable yet.
Support for cfg(doctest) on stable was added in Rust 1.40. A second pinned version aptly named 'pinned_with_doctest'
was added to the CI config. It will use Rust 1.40. As a consequence of the above, the (originally) pinned build will
no longer run doc tests.