Fix truncate_iterator_to_cursor does not respect advance_cursor
The implementation previously did manage the queue, but did not advance the iterator itself. Possibly it can be optimized a bit by using iterator.nth(n) if n > 0, but since I've not ran a benchmark, I've left it out for now.

An even simpler implementation for truncate_iterator_to_cursor could also just not bother with playing around with the internals and just do:

```rust
for _ in 0..self.cursor {
  let _ = self.next();
}
```

Here next is our implementation of the next method for the iterator trait, which would handle the emptying of the queue, albeit one by one instead of in one go. I don't know whether the re-allocation of the queue or emptying it one by one would be faster (it probably depends).